[wip] timing: Add a call graph view using Zest
This is a small prototype (not finished yet) that creates a list of all
the functions called in the call stack, and each function has a link to
all her callees and callers.

This list is created in the CallGraphAnalysis, once the analysis
finishes its execution, the view takes this list and displays it using
the Zest library elements.

The view contains some problems, it doesn't update once the
CallGraphAnalysis finishes its execution. So to try it you need to wait
for the CallGraphAnalysis to finish then open the view.

Change-Id: I6f07d5748896d8f7970f84c67b92d32ca48eca3f
Signed-off-by: Sonia Farrah <sonia.farrah@ericsson.com>