Fix segments aspects issues
Those issues was introduce with the segment filter prototype commit. The
analysis needs to know all the segment aspects so it could rebuild the
filter for any segment aspects. This is a quick temporary solution. The
definitive solution should add some base segment aspects in the Abstract
analysis or in the segment provider interface.

Change-Id: Ie9ad414475d9c84dfa1ae2d82816da1d62ac0440
Signed-off-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>